Minimize Travis CI, Using CodeAnalysis Recommend Fixes Again

+++ src/ICSharpCode.SharpZipLib/Checksum/BZip2Crc.cs
@@ -138,7 +138,7 @@ namespace ICSharpCode.SharpZipLib.Checksum
 				//return (long)(~checkValue ^ crcXor);
 				// but x ^ 0 = x, so there is no point in adding
 				// the XOR operation
-				return (long)(~checkValue);
+				return ~checkValue;
 			}
 		}
